// packages/engine-types/src/models/config/dataConverterConnectorConfig.ts
// Copyright 2024 IOTA Stiftung.
// SPDX-License-Identifier: Apache-2.0.
import type { DataConverterConnectorType } from "../types/dataConverterConnectorType.js";

/**
 * Data converter connector config types.
 */
export type DataConverterConnectorConfig =
	| {
			type: typeof DataConverterConnectorType.Json;
			options?: never;
	  }
	| {
			type: typeof DataConverterConnectorType.Xml;
			options?: never;
	  };